Restructured into controller, views (see commit description)
controller.js is the main flow of the game. this was created to avoid
"chaining" - i.e. each part of the game calling the next part. Instead,
we have the flow of the game in one file.

Furthermore, all jquery/front-end code is moved to respective view
files, which contain both the document.ready selectors (to enable click
actions, etc) as well as a draw() method which draws the canvas or html
for that screen.